refactor : dto 보단 view 의 성향이 강해 view 로 변경
