+++ src/main/kotlin/blackjack/PlayerVo.kt
@@ -1,0 +1,3 @@
+package blackjack
+
+class PlayerVo(val name: String, val cards: String)
